requirements.txt
DateTime==5.5
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
tornado==6.5.1
zope.interface==7.2
SQLAlchemy>=2.0,<3.0
asyncpg>=0.30,<1.0
alembic>=1.13,<2.0
greenlet>=3.0,<4.0
fastapi>=0.100
pydantic>=2.0
redis>=5.0
arq>=0.25
jinja2>=3.1
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation-asgi>=0.48b0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-sqlalchemy>=0.48b0

# Observability
python-json-logger>=2.0.7
prometheus-client>=0.20.0

# Contract testing
schemathesis>=3.26.0

# Testing
pytest>=8.3
pytest-asyncio>=0.23
testcontainers>=4.7.2
